+++ src/editors/LevelEditor/Tools/Base/ESceneCustomOToolsIO.cpp
@@ -26,7 +26,7 @@ bool ESceneCustomOTool::LoadSelection(IReader &F)
     int count = 0;
     F.r_chunk(CHUNK_OBJECT_COUNT, &count);
 
-    SPBItem *pb = UI->ProgressStart(count, make_string("Loading %s(stream)...", ClassDesc()).c_str());
+    SPBItem *pb = UI->ProgressStart(static_cast<float>(count), make_string("Loading %s(stream)...", ClassDesc()).c_str());
     Scene->ReadObjectsStream(F, CHUNK_OBJECTS, EScene::TAppendObject(this, &ESceneCustomOTool::OnLoadSelectionAppendObject), pb);
     UI->ProgressEnd(pb);
 
@@ -57,7 +57,7 @@ bool ESceneCustomOTool::LoadLTX(CInifile &ini)
     inherited::LoadLTX(ini);
 
     u32 count = ini.r_u32("main", "objects_count");
-    SPBItem *pb = UI->ProgressStart(count, make_string("Loading %s(ltx)...", ClassDesc()).c_str());
+    SPBItem *pb = UI->ProgressStart(static_cast<float>(count), make_string("Loading %s(ltx)...", ClassDesc()).c_str());
 
     u32 i = 0;
     string128 buff;
@@ -87,7 +87,7 @@ bool ESceneCustomOTool::LoadStream(IReader &F)
     int count = 0;
     F.r_chunk(CHUNK_OBJECT_COUNT, &count);
 
-    SPBItem *pb = UI->ProgressStart(count, make_string("Loading %s...", ClassDesc()).c_str());
+    SPBItem *pb = UI->ProgressStart(static_cast<float>(count), make_string("Loading %s...", ClassDesc()).c_str());
     Scene->ReadObjectsStream(F, CHUNK_OBJECTS, EScene::TAppendObject(this, &ESceneCustomOTool::OnLoadAppendObject), pb);
     UI->ProgressEnd(pb);
